- oooooh man
git-svn-id: http://svn.codergrid.de/svn/sf/bruteripper/trunk@206 7add9c9f-332d-0410-adcc-f7b1da7091d4

+++ src/org/esb/io/ObjectOutputStream.h
@@ -25,15 +25,12 @@ class ObjectOutputStream:public OutputStream{
 	    boost::archive::text_oarchive archive(archive_stream);
 	    archive << object;
 	    std::string _outbound_data = archive_stream.str();
-//	    int64_t * length=new int64_t;
 	    int length=_outbound_data.length();
-	    
-	    _os->write((char*)&length,sizeof(int));
+	    _os->write((char*)&length,sizeof(int64_t));
 	    _os->write((char*)_outbound_data.c_str(),_outbound_data.length());
 	}
     private:
 	OutputStream * _os;
-//	std::string _outbound_data;
 };
 }}}
 #endif
